Question_1/requirements.txt
# General dependencies
python-dotenv
pandas
numpy
logging

# Azure Storage dependencies
azure-storage-file-datalake

# PostgreSQL dependencies
psycopg2-binary
sqlalchemy

# Airflow dependencies (if used in Airflow DAGs)
apache-airflow